refactor async logout to regular dispatch
